Move `Resolver` to `@ember/-internals/owner`
This enables it to be re-exported correctly from `@ember/owner` with
reference *only* to `@ember/-internals/owner`, while still providing
it to the other `-internals` packages, which should never depend on
the public packages like `@ember/owner`.

In support of this (and many other changes), introduce a type helper
for `FullName` types.