Rename `Owner` to `InternalOwner` in `-internals`
As described in an earlier commit, our public type for `Owner`, as
specified by RFC 0821, is substantially narrower than the existing type
we use as `Owner` internally: our *internal* notion of `Owner` is
the union of `RegistryProxyMixin` and `ContainerProxyMixin`. The new
`InternalOwner` type represents that union.

(We intend, eventually, to carve away the additional parts of
`InternalOwner` until it *is* just the specified public API for
`Owner`, since that is sufficient for everything we actually do now.)

In cases where the consumer does not need `InternalOwner`, switch to
using this new `Owner` default export.

As a knock-on, remove an `any` and actually specify a type which *also*
uses `InternalOwner` internally, which involves an otherwise unrelated
update to type imports and declarations.