Introduce/update type tests for `Owner` internals
For the moment, these aim to match the type tests for the public
preview types for `@ember/owner` as much as possible, but with special
cases covered for the bits which are actually internal. This makes for
a stronger guarantee that we will not have a breaking change for end
users in the future. However, the tests for `@ember/-internals/owner`
*over*-reproduce those, since they also exist for `@ember/owner`
itself. A future commit will therefore remove all of the bits which are
*not* specifically internal for that type test package.