Move `ContainerProxy` and `RegistryProxy` to `@ember/-internals/owner`
These two types are fundamental to the full public notion of an "owner"
in Ember: an owner *in practice* is the combination of these two types,
as when defining `EngineInstance` by applying the two mixins which they
correspond to. Accordingly, the definition of `InternalOwner` *must*
use them... but these internal packages must also use types from the
`@ember/-internals/owner` package to define their own contract; and
moreover those types are *also* part of the `Owner` API contract.

Accordingly, move the types to be internal to `@ember/-internals/owner`
and define the mixins as simply *implementing* these interfaces. (Note:
because these are mixins, the type for "implementing" the interface is
defining a new interface which `extends` the interface, as usual with
our mixin system.)

Here, we define them in terms of extensions of`Owner`, and then define
`InternalOwner` in terms of these two APIs. This needs further tweaks,
as made clear by the need to use `Pick` in one of the definitions and
the fact that both APIs `extends Owner` even though that isn't actually
quite right; a future commit will re-align them to the correct API.