`Factory.create()` cannot be type safe
This means our `.create()` type does not check against the types of the
target object in any way, unfortunately. However, we actually *cannot*
constrain it further than this without going down a *very* deep rabbit
hole (see the historic types for `.create()` on DefinitelyTyped if
you're curious), because we need (for historical reasons) to support
classes which implement this contract to be able to provide a
*narrower* interface than "exactly the public fields on the class"
while still falling back to the "exactly the public fields on the
class" for the general case. Unfortunately, Ember internally relies on
the ability for items which implement `Factory` to be able to customize
the arguments to the `.create()` call in just that way.