Disable a type safety/correctness assertion
In `CoreObject.create()`, we currently have an unsafe cast when
constructing an instance with props, because the code assumes one of
two things. Either: the `factory` is always set, *or* it is safe to
call `setFactoryFor` with `undefined`. A number of acceptance tests are
(incorrectly? Unclear!) relying on the ability to run through this path
with `factory` being `undefined`. It's *possible* that actually means
that the type for `setFactoryFor` should allow `undefined`, but we
typed it the other way for good reason! Accordingly, this *casts*
`factory`, and the commented-out `assert()` is here in the hope that we
can enable it after addressing tests in the future.