Disable another type safety/correctness assertion
The `generateControllerFactory()` function *should* require that the
class associated with the `FactoryManager` for a given controller name
be a `Controller` subclass, but the tests today do *not* keep this
contract. We should fix that, but *separately* from this types-focused
change, which aims to change *only* the types and module graph.