Refactor tsconfigs to share compiler options
This lets each config specify *only* how it actually does a build (or
not!), while sharing the config explicitly.

It also fixes an existing bug in the compilation settings which was not
affecting *normal* TS compilation with `tsc`, but was generating noise
in the `broccoli-typescript-compiler` pipeline because the packages
happened to be resolved in a different order such that the `loader`
"package" definition, and its declaration of the `require` module, was
not present in the graph. Set it explicitly in `compilerOptiosn.paths`
to fix that.