Introduce a script to publish stable types
Provide a script which runs the compiler against a new tsconfig for
generating types, wraps all the generated modules in `declare module`
statements, and then creates an `index.d.ts` which uses side-effect
style imports to expose them all, just the same as the preview types
but generated from source.

Critically, this infrastructure does not expose *any* stable types in
and of itself. Instead, it introduces a list of all the types still in
preview mode which acts as a filter, and currently *all* modules are in
that filter. Stabilizing the types for a given module will mean
removing modules from that list and removing the corresponding preview
types definitions.