Update to TS v4.9
This lets us take advantage of much smarter `in`-based type narrowing
and the new `satisfies` operator internally. We'll need to make a
decision about what versions to support when we actually "stabilize"
the types: not merely publishing them, but opting consumers into them
by default. However, it's likely that 4.9 and 5.0 will be reasonable
targets at that point, and even if not the particular 4.9 features we
might rely on are "runtime" TS features, so we should be able to add a
CI check for 4.8 if we want to add support for it in the test matrix.