Remove runtime dependence on `InternalOwner`
As a follow-on to #20310, this removes all remaining direct uses of the
`InternalOwner` type; the only place we use it now is in the defintion
of `EngineInstance`. This frees us to start chipping away further at
this type: we can deprecate *nearly all of it*, and we know that our
internal code does not depend on any of it except where it explicitly
uses `EngineInstance`.

Part of #20303.