Use `Engine` API in `verifyRegistration`
This is a dumb set of tests, which are verifying internal mechanics of
*how* we make this work, rather than verify *that* it works for each of
these issues. However, for the moment the key is to unblock decoupling
work, so I'm creating [a separate issue][1] to make sure we cover this
in other ways and then delete the tests which cause this coupling.

[1]: https://github.com/emberjs/ember.js/issues/20315