Normalize to always import from `'backburner.js'`
We had a mix of different import locations for this, both the ES
imports via `backburner.js` and the Ember CLI-munged version available
at `backburner`. Switch to only using the Node-resolveable version, so
TS can resolve it correctly; and enforce it by removing `backburner`
from the list of `core` modules in the ESLint config.