types: work around backburner runtime paths shenanigans
The legal runtime import for the `backburner.js` package is apparently
`backburner`, but I have not yet been able to identify where that
rename occurs. We should ultimately align that to use the package name
instead of a custom import specifier, but for now, unblock publishing
types from source by rewriting all imports in the emitted types from
`'backburner'` to `'backburner.js'`.