Change how loader's `define` becomes visiblte to TS
Previously, this was defined as a global in `ember-template-compiler`.
This was not really correct; it was a workaround for the way that the
`loader` package defines it as a global as a side effect of importing
and executing the module. Here, make it visible specifically *and only*
to the root `ember` package, where it is needed for defining the type
of the `Ember.__loader` object.