Provide an `owner-ext` for `@ember/service`
`owner-ext` is a pass-through type definition module which adds the
`Registry` from `@ember/service` to the `DIRegistry` in `@ember/owner`
in a way that works for both internal usage within Ember *and* usage by
end users, given the constraints of our types publishing infrastructure
and its need to wrap all proper modules in `declare module { }` blocks.
Since this already *has* a `declare module { }` block, it does not need
to be wrapped; and since it is simply a `.d.ts` file, we can copy it
over rather than trying to emit it with `tsc` during build. (This is,
admittedly, a hack: but a sadly necessary one given our constraints!)