[FEATURE] generate type registry entry in service blueprint
We have support for this capability in the types, so the blueprints
should make sure that there is a registry entry for each service. We
originally left this out because we were unsure of the future of the
service registry, but we made `Owner.lookup()` smart enough to resolve
it, and the `@service` decorator continues to expect it.