Remove `escapeExpression` from `@ember/template`
This is unused elsewhere in the framework itself and is not public API,
so it should be safe to remove. Additionally, since this is just a copy
of the API from within Handlebars itself, users who want to use it
should directly use Handlebars' (or some other library's) escaping
library instead. See [the discussion at #16817][16817] for background.

[16817]: https://github.com/emberjs/ember.js/issues/16817