[CLEANUP] Remove deprecated `AutoLocation` class
Replace `Location` object with an interface. We no longer need the
object (it does not do anything) and in fact exporting it would just
troll people. However, we *do* need the interface internally, and it is
useful to allow people to name the 'any Location type' type, so export
that in its place.

Along the way, delete a test which prevented the DI registry type from
working correctly in Ember's internals, and simplify a few of the type
lookups there.