[BUGFIX LTS] Fix runloop types on TS 5.0+
TS 5.0 catches a new set of assignability errors here, and our (very)
complicated set of types for `bind` and friends got flagged by that.
The terminal case for `PartialParams` needs to be a fixed empty tuple
rather than `never`, so that the type system can see that it will be
resolved as 'no arguments' rather than 'nothing' in that scenario.