Publish types from source and remove preview types
Remove *all* preview types, update the publishing script to publish all
packages from source, and update the publishing tsconfig to include all
source packages.

**This *does not work* yet.** However, it is a necessary first step in
the process of switching to publishing stable types: the remaining
types are too deeply entangled with each other to publish them in
chunks as we have to date. (If there are exceptions, I have not found
them, and the overhead of finding them is higher than the cost of just
doing this switch-over.)