Add ability to manually copy `.d.ts` type modules
Some packages in the project supply `.d.ts` files which must be part of
both the authoring-time types *and* the published types, because they
represent values which become available at runtime by way of codegen.

For example, the `ember/version` module provides a `VERSION` value
derived from the package version. Both Ember's internals and the public
API include this value in various ways, but it does not have a
corresponding ES module on disk, so we provide a type definition
representing where it *notionally* lives, and that type definition must
also appear in the published types.

In these cases, simply copy the file from an input to an output
location, rather than trying to use `tsc` for them (since `tsc` will
actually ignore them).