Handle type (un)safety of some runloop `bind`
We can safely type *almost* all variants, but not using Ember's `bind`
with `this` and a string name. Introduce an appropriate fallback for
that case; this means that we cannot catch quite as many errors but in
exchange we get coverage over documented patterns in Ember's code.

(This gives us the same behavior as the preview and DT types always had
as a fallback.)