Refactor `Ember` namespace to be a TS `namespace`
While the `namespace` construct is very much to be avoided for new
code, it is *exactly* correct for the case of the `Ember` namespace,
which is (precisely!) a namespace. To make this work:

- Combine the existing `const PartialEmber` and `const Ember` object
  bindings into a single `namespace Ember`.

- For the members assigned via `Object.defineProperty`, switch from an
  `interface` to `export declare let` and `export declare const`
  bindings on the `namespace`. Make sure to use `let` or `const` to
  match the readonly or writable semantics of the items defined thus.

Testing this exposed that at some point `Ember.run.later` and similar
nested namespace access stopped working, almost certainly at the point
when we dropped support for the namespace access from the module, but
our type tests still promised it. Noting that no *runtime* tests check
for this, remove the corresponding type tests.