Simplify types for `(get|set)Properties`
This fixes many of the incompatibilities from our type tests, and makes
the resulting code *much* clearer at the same time. The key here is to
move away from trying to use tuple spread types to just constraining
the actual args and mapping their types into a `Pick` or `Record`.