Eliminate cycle in definition of (Ember) `Component`
`Component` uses `ViewSupport`, but `ViewSupport` defined its `this` as
`Component`. This was a classic form of cycle which blocked supporting
the `Mixin`-based programming model effectively with TS in the first
place and was a major motivator for investing in native classes back in
the early Ember 3.x days! here, the fix is to use `this: CoreView`
instead so that `ViewSupport` does not need to name `Component`, only
the chunk of `Component`'s base types it actually uses.

As a bonus, remove a needless `extends FrameworkObject` from the merge
on the interface of the `CoreView` class, since normal inheritance does
correctly preserve that.