Fix/simplify types of `create`, `reopen`, and `reopenClass`
As described in the RFC, we do not want to attempt to provide the type-
modifying capabilities of these functions. The result is type safe but
not especially useful, but that is as we want: using `create` to
provide values not on the base type *or* to change its values is quite
undesirable; and using `reopen` and `reopenClass` to dynamically update
instance and static members of a class has serious (and not obvious!)
performance downsides and discoverability problems for maintainers, and
will ultimately be deprecated entirely as part of removing Classic.

A key result is that `.create()` now allows *any* props to be set, but
does not update the resulting type. This is an unfortunate necessity
given how Ember *itself* uses `.create()`.