Partially reimplement type-safe `.create()`
This gets us a *small* amount of type safety, rather than absolutely
none: it does not check the types passed into `.create()` match, but
since there is little reason to directly call this for end user code
(vs. for internal code, which actually uses this dynamic capability!),
this gives us the actual runtime behavior: if you smash a property with
a new type, you... do.