Update a `set()` type test: it allows dynamic setting
This matches the runtime behavior. We do not catch setting things this
way, as a result, but in a world where native property access works
everywhere, this is actually somewhat useful: it means you can set
anything on a target object, and although the type will not be updated
with that change, it will at least type check in a way that matches the
dynamic runtime behavior.