Make Ember's helper types Glint-compatible
Provide the same public API as the preview types. This requires lying a
bit about the return type of `helper()` in particular, but no part of
Ember's own internals cares about the details: they are relevant *only*
as the point where they are actually created. The provided type avoids
a number of possible kinds of errors by providing an opaque item which
cannot be constructed but which *does* preserve generics. Moreover, the
actual details of type returned by `helper()` are *not* public API, so
this correctly avoids exposing those details.

As a bonus, remove a number of extra internal type aliases on both the
source and that were used in the preview types.