Publicly export `FunctionBasedHelperInstance`
This abstract class -- specifically, its `protected abstract
__concrete__` member -- prevents subclasses from doing `class X extends
helper(..)`, since that is an error at runtime. While it is rare that
people would type that, it is not impossible and we use this to give
them early signal via the types for a behavior which will break (and in
a somewhat inscrutable way!) at runtime.

This is needful because we lie about what this actually is for Glint's
sake: a function-based helper returns a `Factory<SimpleHelper>`, which
is designed to be "opaque" from a consumer's POV, i.e. not
user-callable or constructible but only useable in a template (or via
`invokeHelper()` which also treats it as a fully opaque `object` from a
type POV). But Glint needs a `Helper<S>` to make it work the same way
as class-based helpers. (Note that this does not hold for plain
functions as helpers, which it can handle distinctly.) This signature
thus makes it so that the item is usable *as* a `Helper` in Glint, but
without letting end users treat it as a helper class instance.

Long-term, this entire construct can go away in favor of deprecating
the `helper()` invocation in favor of using plain functions.