Restore performance tests (#1427)
This PR was originally intended to reactivate the performance tests. My plan was to run the tests through Vite, which revealed more inconsistencies with our module setup.

This PR continues #1426 by tightening up the lints, adding a number of lints to catch problems that were stopping the built modules from working in the vite setup in the performance suite. All of the lints now run on all packages (previously, the non-published packages had significantly weaker lints), and both the published and non-published packages now share an in-repo linting plugin to keep things consistent.

It also tightens up and solidifies the infrastructure that built the packages and .d.ts files.

There's still a bit more to do to finalize the exports setup that will be driven by finishing up the performance suite work.

Since these changes are standalone improvements, I'm going to land this with the perf suite still disabled.