[BUGFIX LTS] Fix cyclic module crash
`@ember/engine/index` and `@ember/engine/instance` have circular dependencies on each other.

This would be OK, except that `instance.ts` is eagerly copying a class off `index.ts`, at a point in time where it might not actually be available.

Whether this bug hits you depends on which side of the pair gets evaluated first. I only happened hit it when testing an app with engines and fastboot on Ember 5.

This fix is to not eagerly pull the class off the module.

(cherry picked from commit 7c31adae643dbc209713cb998e77f541f50b57db)