Upgrade TS and configure TS CI accordingly
- Bump the development version to 5.1.
- Update the CI config to test against the other supported versions:
  currently 4.9 and 5.0, which we tested against implicitly previously
  by way of our tests against TS nightly releases.

This is a bit more important now that we are officially publishing
stable types! Using 4.9 as our initially supported version is an easy
choice: it is what we were already using, and it gives us a solid six
months of TS releases as our initial support category. Folks may find
it useful to review [the "rolling windows" policy][semver-ts-rw], since
that is Ember's SemVer policy for TS.

[semver-ts-rw]: https://www.semver-ts.org/#rolling-support-windows