[BUGFIX stable] Fix type dependency issues
Add `@simple-dom/interface` and `@glimmer/syntax` to `dependencies` so
they are present in the graph for consumers who will not be getting
them transitively from the `@glimmer/*` deps which are pre-built into
Ember. (Medium-term, it would be preferable for our type emit to roll
up these kinds of dependencies the same way Ember's build does *or* for
Ember's build to stop rolling them up; whichever we choose, the goal
will be for them to *match*.)

Fixes #20479.