[BUGFIX stable] Fix types for Resolver contract
We need `Resolver` implementors (e.g. `ember-resolver`) to be able to
provide a class which can satisfy this contract, which means we need to
support either a static `.create()` method or a constructor (with a
strong preference for the constructor). This is likely to be a common
issue with this kind of assignability of classic classes, because it
ultimately comes down to the definition of `Factory` and `CoreObject`.

To resolve this, add two overloads for `CoreObject.prototype.create`:

- One which just produces an instance type with no args at all.
- One which produces an instance type by merging with passed-in args.

These two overloads also improve the type safety of `.create()` more
generally, allowing us to handle a couple cases which the types did not
previously account for (and indeed had given up handling).

This also adds a smoke test for something shaped like the `app.ts` file
generated for every Ember app, so that we can be sure that the code we
emit when generating an app actually works. For now, this simply copies
over the types from `ember-resolver`, `ember-load-initializers`, and a
basic `environment`.