[BUGFIX stable] Fix model arg to a Route `serialize`
Previously, the internal type here always explicitly added `| undefined`
to the type of the `model` param. This was likely reasonable when the
conversion first happened, since the type here was simply a minimal
`IModel | undefined`. Since the param is now driven by the type param
for the `Route`, though, and defaults to `unknown`, we can just use the
`Model` type param directly, and the implementation handles it correctly
since the was actually already, albeit invisibly, just `unknown`, since
`unknown | undefined` === `unknown`.