Update ApplicationInstance#visit to use followRedirects()
The current implementation of `Transition#followRedirects()` is identical to the logic here: it catches rejections, and then checks for any other `activeTransition`. Therefore, this commit introduces no change in behavior.

But, if/when https://github.com/tildeio/router.js/pull/335 lands, it will means we can take advantage of that fix for ApplicationInstance#visit.