[BUGFIX canary] Restore/deprecate outlet `TemplateFactory` support
Fixes #20576

The test is a bit misleading in that, this is not actually restored
to support the very old 1.6.0 version of ember-test-helpers. The
issue is that while that particular spot of the invalid usage was
fixed in 1.6.1, other similarly invalid patterns have popped up in
the code since then. However, the old tests was probably a close
enough approximation of the general problem.

This is not to avoid fixing the actual issue, but it's just that it
was very difficult for us to support for a short time, and there
may be other code that copied the same invalid pattern elsewhere.

See also https://github.com/emberjs/ember-test-helpers/pull/1445