[CLEANUP canary] Enable tsconfig isolatedModules
This enables typescript's [isolatedModules](https://www.typescriptlang.org/tsconfig#isolatedModules) flag. It forces our source to be clearer about what is a type vs a value.

This is a step toward a reformed build for ember-source. isolatedModules means we can use babel to handle the typescript instead of needing a global tsc step. This PR doesn't make that switch, that will come as a followup.

I removed `@types/node` because it doesn't typecheck under isolatedModules. We were only using two tiny things out of it, which were easy enough to declare at point of use.