Improve link script for glimmer-vm packages
Since packages/** are real packages now, it is not sufficient to
only link from the root (e.g. for TypeScript checking), and instead
we will need to link each sub-package separately.

In the future, https://github.com/pnpm/rfcs/blob/main/text/0001-catalogs.md
may help, but that is not yet implemented.