[BUGFIX beta] internal templates should be strictMode
As of https://github.com/emberjs/ember.js/pull/20587, the modules published under `dist/packages` act more like normal addon code,  in that their templates are published as calls to `precompileTemplate` rather than wire format.

But that makes this the first time that Embroider is seeing these templates, and because they are in loose mode and use the dynamic component helper they are un-analyzable.

This PR:
 - switches them to strict mode
 - fixes the declared types for `precompileTemplate`, which were a lie before. `scope` is allowed in non-strict mode. `scope` is optional even in strict mode.
 - fixes a spelling error in a recently-introduced internal method, because I noticed it and it annoyed me.