[failing test] add test for adding/removing/re-adding modifier with `if`
This test currently fails if the modifier is inside an `(if)` statement,
and if the condition changes from true -> false -> true.

It seems like somehow when the 3rd re-render to true happens, the `modifier`
helper gets a clone of the first two positional arguments (the event name and the
callback). The `modifier` helper gets 4 arguments when it should only be getting two.