Revert "[perf] replace Math.max usage to reduce and pure compare"
This reverts commit 7a86ffca3ec66b1c1014c4caa2afc15d6d43cc7f.

`Math.max()` and `>` have different reuslts for `NaN` which is used
by volatile tag. This is caught by tests but wasn't reported in the
harness correctly.

I suspect we can drop volatile tags, I am pretty sure Ember doesn't
use it anymore. However that would require some work to clean up
probably.

The original commit was merged permaturely anyway. We may be able
to re-land a targeted version of that commit, with better testing,
once/if volatile tags are removed.