[BUGFIX] properly surface deprecation message
I think wrapping these in LOCAL_DEBUG ends up causing these to not
actually be seen by anyone. It's also unnecessary, `deprecate`
should already be pluggable/strippable in Ember, but it probably
doesn't matter in this case anyway, because most of the time the
compiler isn't shipped to the browser anyway and should be (???)
running in dev mode by default? I hope?

Anyway if that last part isn't true and these still aren't being
seen, then it would be something to fix in Ember.