Deprecate `Program` in v1 AST
`{ type: "Program" }` isn't a real AST node you actually encounter,
it was inherited from the Handlebars AST, but it has been split
into AST.Template and AST.Block in Glimmer for a long time now.

More throughly deprecate the `Program` visitor and builders.

Also rename `blockParams` on `Template` to `locals`, and make both
the array and the property immutable. We are only providing this
information so that AST consumers can make decisions on what names
are referring to outer local variables, but the compiler does not
refer to it at all, and always use the original `options.locals`
as the source of truth, so if any AST plugin tried to mutate this,
it would not have worked.

Keeping `blockParams` around as a deprecated alias for the time
being. Consumers that need to support older versions should be able
to feature detect this (such as with `"locals" in node`, or even
just short-circuiting `node.locals ?? node.blockParams` to avoid
triggering the deprecation warning.