Fix types for ASTv1 "callable" nodes
Other than mustache, none of the other "callable" nodes allow a
literal in the "callee" (node.path) position:

{{log ("hi")}}
       ~~~~ illegal

<div {{"hi"}} />
       ~~~~ illegal

{{#"hi"}}...{{/"hi"}}
   ~~~~ illegal

Further more, in mustache positions this is only allowed if there
are no other arguments:

{{"hi" 1 2 3}}
  ~~~~ illlegal

Most of these syntaxes already generate syntax errors in the parser
but is allowed in the v1 AST, so it's possible for a AST plugin to
force these illegal conditions in a transform but should ultimately
get caught and rejected elsewhere.

This commit fixes the type to reflect that these conditions are
illegal/impossible and make sure we catch all of these edge cases
in the v2 normalization step.