Deprecate AST v1 literal.original
These probably once existed to normalize across all the possible
nodes that call go into a `mustache.path` position, such that you
can always do `node.path.original` on a mustache.

In my opinion, this is of dubious utility in the first place – the
`original` property on these can have a variety of different types,
it's unclear what you would do with the generic value you acquired
in this manner. Code like `if (node.path.original === "component")`
will likely misclassify `{{"component"}}` as `{{component}}`.

Further more, with the introduction of the `{{(sexp)}}` syntax, the
original assumption no longer holds anyway, as sub-expression nodes
do not have an `original` property.

Deprecating this property and turning it into a getter/setter for
`value`, ensuring they stay in sync.

In the future, we may want to reclaim this original property to
store the _actual_ original syntax in the form of a string. For
example, today, if you had `{{1.0}}` in the source code, it gets
parsed into the JavaScript number/value `1` and the original "raw"
format is irrecoverably lost, which is a problem for printer and
codemods type use cases.