Funnel all v1 AST node construction through parser builders
The parser builders are the defacto constructors for the v1 AST
nodes. As the nodes get increasingly more eloborate for backwards
compatibility and such, it's important that we share the logic for
constructing these nodes as much as possible so we don't miss any
cases.

Also standardize the signatures to use the object argument style,
you should be able to clone a node by passing it back into the same
builder method which it originated from.