Second attempt at landing #1552 (part 2/2)
The previous/existing "parsing" code tries to piggy back on letting
the HTML tokenizer/parser parse out the block params syntax into
attribute nodes and tries to recover the information after the fact

This is fundamentally broken and suffers from some wild edge cases
like:

```
<div as="" |foo="bar|>...</div> // => <div as |foo|>...</div>
```

This is especially problematic in light of wanting to retain the
source spans of the block params nodes.

This commit rework the block params parsing and integrate it into
the normal parsing pass in the appropiate time.